tests/wpt/web-platform-tests/speculation-rules/prerender/activation-start.html

<!DOCTYPE html>
<title>PerformanceNavigationTiming.activationStart in prerendered page</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

setup(() => assertSpeculationRulesIsSupported());
promise_test(async t => {
  const uid = token();
  const testChannel = new PrerenderChannel('test-channel', uid);
  t.add_cleanup(() => {
    testChannel.close();
  });
  const gotMessage = new Promise(resolve => {
    testChannel.addEventListener('message', e => resolve(e.data), {once: true});
  });
  window.open(`resources/activation-start.html?uid=${uid}`, '_blank', 'noopener');
  assert_equals(await gotMessage, 'Done');
}, 'PerformanceNavigationTiming.activationStart in prerendered page');

</script>
</body>
